+++ lib/msf/core/post/windows/reflective_dll_injection.rb
@@ -1,5 +1,7 @@
 # -*- coding: binary -*-
 
+require 'msf/core/reflective_dll_loader'
+
 ###
 #
 # This module exposes functionality which makes it easier to do
@@ -7,7 +9,9 @@
 #
 ###
 
-module Msf::ReflectiveDLLInjection
+module Msf::Post::Windows::ReflectiveDLLInjection
+
+  include Msf::ReflectiveDLLLoader
 
   PAGE_ALIGN = 1024
 
@@ -32,33 +36,6 @@ module Msf::ReflectiveDLLInjection
     process.memory.write(shellcode_mem, shellcode)
 
     return shellcode_mem
-  end
-
-  #
-  # Load a reflectively-injectable DLL from disk and find the offset
-  # to the ReflectiveLoader function inside the DLL.
-  #
-  # @param dll_path Path to the DLL to load.
-  #
-  # @return [Array] Tuple of DLL contents and offset to the
-  #                 +ReflectiveLoader+ function within the DLL.
-  #
-  def load_rdi_dll(dll_path)
-    dll = ''
-    offset = nil
-
-    ::File.open(dll_path, 'rb') { |f| dll = f.read }
-
-    pe = Rex::PeParsey::Pe.new(Rex::ImageSource::Memory.new(dll))
-
-    pe.exports.entries.each do |e|
-      if e.name =~ /^\S*ReflectiveLoader\S*/
-        offset = pe.rva_to_file_offset(e.rva)
-        break
-      end
-    end
-
-    return dll, offset
   end
 
   #
