+++ modules/exploits/windows/local/ms10_015_kitrap0d.rb
@@ -4,16 +4,16 @@
 ##
 
 require 'msf/core'
-require 'msf/core/reflective_dll_injection'
+require 'msf/core/post/windows/reflective_dll_injection'
 require 'msf/core/exploit/exe'
 require 'rex'
 
 class Metasploit3 < Msf::Exploit::Local
   Rank = GreatRanking
 
-  include Msf::ReflectiveDLLInjection
-  include Post::File
-  include Post::Windows::Priv
+  include Msf::Post::File
+  include Msf::Post::Windows::Priv
+  include Msf::Post::Windows::ReflectiveDLLInjection
 
   def initialize(info={})
     super( update_info( info,
