+++ modules/exploits/windows/local/ppr_flatten_rec.rb
@@ -4,17 +4,17 @@
 ##
 
 require 'msf/core'
-require 'msf/core/reflective_dll_injection'
+require 'msf/core/post/windows/reflective_dll_injection'
 require 'rex'
 
 class Metasploit3 < Msf::Exploit::Local
   Rank = AverageRanking
 
-  include Msf::ReflectiveDLLInjection
   include Msf::Post::File
   include Msf::Post::Windows::Priv
   include Msf::Post::Windows::Process
   include Msf::Post::Windows::FileInfo
+  include Msf::Post::Windows::ReflectiveDLLInjection
 
   def initialize(info={})
     super(update_info(info, {
